+++ spec/pacto/pacto_spec.rb
@@ -1,5 +1,4 @@
 describe Pacto do
-  let(:host) { 'http://localhost' }
   let(:contract_name) { 'contract' }
   let(:contract) { double('contract') }
 
@@ -24,19 +23,19 @@ describe Pacto do
   end
 
   describe '.build_from_file' do
-    let(:contract_path) { File.join('spec', 'data', "#{contract_name}.json") }
-    let(:file_pre_processor) { double('file_pre_processor') }
-    let(:file_content) {File.read(contract_path)}
-    
-    it 'should build a contract given a JSON file path and a host' do
-      file_pre_processor.stub(:process).and_return(file_content)
-      described_class.build_from_file(contract_path, host, file_pre_processor).
-        should be_a_kind_of(Pacto::Contract)
+    let(:path)                  { 'contract/path' }
+    let(:host)                  { 'http://localhost' }
+    let(:file_pre_processor)    { double('file_pre_processor') }
+    let(:instantiated_contract) { double(:instantiated_contract) }
+
+    it 'delegates to ContractFactory' do
+      Pacto::ContractFactory.should_receive(:build_from_file).with(path, host, file_pre_processor)
+      described_class.build_from_file(path, host, file_pre_processor)
     end
-    
-    it 'should process files using File Pre Processor module' do
-      file_pre_processor.should_receive(:process).with(file_content).and_return(file_content)
-      described_class.build_from_file(contract_path, host, file_pre_processor)
+
+    it 'returns whatever the factory returns' do
+      Pacto::ContractFactory.stub(:build_from_file => instantiated_contract)
+      described_class.build_from_file(path, host, file_pre_processor).should == instantiated_contract
     end
   end
